GolombBuffer::GetPos() return int value. removed the unnecessary conversion. fixed some warnings.
git-svn-id: https://mpc-hc.svn.sourceforge.net/svnroot/mpc-hc/trunk@4905 10f7b99b-c216-0410-bff0-8a66a9350fd8

+++ src/DSUtil/GolombBuffer.cpp
@@ -79,7 +79,7 @@ void CGolombBuffer::BitByteAlign()
 	m_bitlen &= ~7;
 }
 
-__int64 CGolombBuffer::GetPos()
+int CGolombBuffer::GetPos()
 {
 	return m_nBitPos - (m_bitlen>>3);
 }
